@@ -1,31 +1,32 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { BehaviorSubject } from 'rxjs';
+import { BehaviorSubject, Observable, of } from 'rxjs';
 import { catchError, tap } from 'rxjs/operators';
+import { OlympicCountry } from '../models/Olympic';
 
 @Injectable({
   providedIn: 'root',
 })
 export class OlympicService {
   private olympicUrl = './assets/mock/olympic.json';
-  private olympics$ = new BehaviorSubject<any>(undefined);
+  private olympics$ = new BehaviorSubject<OlympicCountry[] | null>(null); // BehaviorSubject pour stocker les données
 
   constructor(private http: HttpClient) {}
 
-  loadInitialData() {
-    return this.http.get<any>(this.olympicUrl).pipe(
-      tap((value) => this.olympics$.next(value)),
-      catchError((error, caught) => {
-        // TODO: improve error handling
-        console.error(error);
-        // can be useful to end loading state and let the user know something went wrong
-        this.olympics$.next(null);
-        return caught;
+  loadInitialData(): Observable<OlympicCountry[] | null> {
+    return this.http.get<OlympicCountry[]>(this.olympicUrl).pipe(
+      tap((data) => this.olympics$.next(data)), // Met à jour les données du BehaviorSubject
+      catchError((error) => {
+        console.error('Erreur lors du chargement des données :', error);
+        this.olympics$.next(null); // Indique une erreur
+        return of(null);
       })
     );
   }
 
-  getOlympics() {
+  // Méthode pour exposer les données sous forme d'Observable
+  getOlympics(): Observable<OlympicCountry[] | null> {
     return this.olympics$.asObservable();
   }
+
 }
